+++ pages/teach/courses/_code/evaluations/_id/index.vue
@@ -19,16 +19,28 @@
             <v-tabs-items v-model="currentTab">
               <v-tab-item>
                 <!-- Either show comment (published/unpublished) or explanation (requested) -->
-                <div v-if="evaluation?.status !== 'REQUESTED'">
+                <div v-if="showComment">
                   <h4>{{ $t('evaluation.comment._') }}</h4>
 
-                  <description-content :text="evaluation?.comment" />
+                  <description-content
+                    entity="evaluation.comment"
+                    :text="evaluation?.comment"
+                  />
                 </div>
 
-                <div v-else>
+                <div v-if="showRejectionReason">
+                  <h4>{{ $t('evaluation.rejectionReason') }}</h4>
+
+                  <description-content :text="evaluation?.rejectionReason" />
+                </div>
+
+                <div v-if="showExplanation">
                   <h4>{{ $t('evaluation.explanation._') }}</h4>
 
-                  <description-content :text="evaluation?.explanation" />
+                  <description-content
+                    entity="evaluation.explanation"
+                    :text="evaluation?.explanation"
+                  />
                 </div>
 
                 <!-- Show list of competencies and checked items/learning outcomes -->
@@ -42,10 +54,13 @@
                 />
 
                 <!-- Show note (published/unpublished) -->
-                <div v-if="evaluation?.status !== 'REQUESTED'">
+                <div v-if="showNote">
                   <h4>{{ $t('evaluation.note._') }}</h4>
 
-                  <description-content :text="evaluation?.note" />
+                  <description-content
+                    entity="evaluation.note"
+                    :text="evaluation?.note"
+                  />
                 </div>
               </v-tab-item>
 
@@ -55,6 +70,11 @@
             </v-tabs-items>
           </v-card-text>
         </v-card>
+
+        <div v-if="evaluation?.isRequestPending" class="text-right mt-3">
+          <evaluation-request-reject-btn @reject="reject" />
+          <accept-btn @accept="accept" />
+        </div>
       </v-col>
 
       <v-col
@@ -68,6 +88,7 @@
       <actions-menu
         :custom-action="customAction"
         :delete-action="deleteAction"
+        :edit-link="editLink"
         @customActionClicked="publish"
       />
     </v-row>
@@ -77,7 +98,9 @@
 </template>
 
 <script>
+import acceptEvaluationRequest from '~/gql/teach/acceptEvaluationRequest.gql'
 import publishEvaluation from '~/gql/teach/publishEvaluation.gql'
+import rejectEvaluationRequest from '~/gql/teach/rejectEvaluationRequest.gql'
 
 export default {
   name: 'TeachEvaluationPage',
@@ -86,6 +109,7 @@ export default {
       course: null,
       currentTab: 0,
       evaluation: null,
+      requestHandling: false,
       title: '',
     }
   },
@@ -97,7 +121,10 @@ export default {
       return this.$route.params.code
     },
     customAction() {
-      if (!this.evaluation || this.evaluation.status !== 'UNPUBLISHED') {
+      if (
+        !this.evaluation ||
+        !['ACCEPTED', 'UNPUBLISHED'].includes(this.evaluation.status)
+      ) {
         return null
       }
 
@@ -107,7 +134,10 @@ export default {
       }
     },
     deleteAction() {
-      if (!this.evaluation || this.evaluation.status === 'REQUESTED') {
+      if (
+        !this.evaluation ||
+        !['PUBLISHED', 'UNPUBLISHED'].includes(this.evaluation.status)
+      ) {
         return null
       }
 
@@ -123,11 +153,71 @@ export default {
         object: 'evaluation',
       }
     },
+    editLink() {
+      if (
+        !this.evaluation ||
+        !['ACCEPTED', 'UNPUBLISHED'].includes(this.evaluation.status)
+      ) {
+        return null
+      }
+
+      return {
+        name: 'teach-courses-code-evaluations-id-edit',
+        params: { code: this.courseCode, id: this.evaluationId },
+      }
+    },
     evaluationId() {
       return this.$route.params.id
     },
+    showComment() {
+      return ['ACCEPTED', 'PUBLISHED', 'UNPUBLISHED'].includes(
+        this.evaluation?.status
+      )
+    },
+    showExplanation() {
+      return ['ACCEPTED', 'REJECTED', 'REQUESTED'].includes(
+        this.evaluation?.status
+      )
+    },
+    showNote() {
+      return ['ACCEPTED', 'PUBLISHED', 'UNPUBLISHED'].includes(
+        this.evaluation?.status
+      )
+    },
+    showRejectionReason() {
+      return this.evaluation?.status === 'REJECTED'
+    },
   },
   methods: {
+    async accept() {
+      if (this.requestHandling) {
+        return
+      }
+
+      this.requestHandling = true
+      const data = { id: this.evaluationId }
+
+      try {
+        const response = await this.$apollo
+          .mutate({
+            mutation: acceptEvaluationRequest,
+            variables: data,
+          })
+          .then(({ data }) => data && data.acceptEvaluationRequest)
+
+        if (response) {
+          // TODO: replace with cache update.
+          this.evaluation.competencies = null
+          this.evaluation.isRequestPending = false
+          this.evaluation.status = 'ACCEPTED'
+          this.$notificationManager.displaySuccessMessage(
+            this.$t('success.EVALUATION_REQUEST_ACCEPT')
+          )
+        }
+      } catch (err) {}
+
+      this.requestHandling = false
+    },
     async publish() {
       const data = { id: this.evaluationId }
 
@@ -147,6 +237,35 @@ export default {
           )
         }
       } catch (err) {}
+    },
+    async reject(reason) {
+      if (this.requestHandling) {
+        return
+      }
+
+      this.requestHandling = true
+      const data = { id: this.evaluationId, reason }
+
+      try {
+        const response = await this.$apollo
+          .mutate({
+            mutation: rejectEvaluationRequest,
+            variables: data,
+          })
+          .then(({ data }) => data && data.rejectEvaluationRequest)
+
+        if (response) {
+          // TODO: replace with cache update.
+          this.evaluation.competencies = null
+          this.evaluation.isRequestPending = false
+          this.evaluation.status = 'REJECTED'
+          this.$notificationManager.displaySuccessMessage(
+            this.$t('success.EVALUATION_REQUEST_REJECT')
+          )
+        }
+      } catch (err) {}
+
+      this.requestHandling = false
     },
     setResult({ data }) {
       if (!data) {